refactor(android): extract `Keyboard` composable
