+++ liushu-android/app/build.gradle
@@ -39,7 +39,7 @@ android {
         compose true
     }
     composeOptions {
-        kotlinCompilerExtensionVersion '1.4.0'
+        kotlinCompilerExtensionVersion '1.4.8'
     }
     packagingOptions {
         resources {
