+++ liushu-android/build.gradle
@@ -10,7 +10,7 @@ buildscript {
 plugins {
     id 'com.android.application' version '8.0.0' apply false
     id 'com.android.library' version '8.0.0' apply false
-    id 'org.jetbrains.kotlin.android' version '1.8.0' apply false
+    id 'org.jetbrains.kotlin.android' version '1.8.22' apply false
     id "org.mozilla.rust-android-gradle.rust-android" version "0.9.3"
 }
